submissions/littleme/littleme/requirements.txt
google-generativeai
typer
rich